refactor(crates): integrators 彻底去重，spice-gated clippy 清零
去重（承接 #248，双树副本的另一半）：
- 纯数学模块（butcher/pd45/pd78/rk89/abm/cowell/multistep_methods/
  rk_methods/solve_ivp）改用 e2m2e-propagation；propagation 新增
  pyo3 optional feature 承载 RkMethod/MultistepMethod 的 pyclass
  （属性与原 integrators 版逐字一致，Python 枚举名不变）。
- compiled/compiled_stm/srp/relativistic 搬到 e2m2e-forces
  （调用点仅 9 行路径替换）；gravity_field 双实现合并（forces 版
  是逐行超集，纯路径替换）；pole_tide 并入 forces 的 solid_tide。
- 删除 integrators 本地 forces/ 整目录、spice_ffi/spk_accel/
  solid_tide/spherical_harmonic 副本及无人调用的 single_shooting
  模块；修复 pymodule 中 segmented_shooting 的重复注册。
- _integrators 导出清单去重前后逐字一致；服务器 spice 构建下
  全量 pytest 1401 passed 与基线持平。

lint：清零 spice-gated 代码全部 16 个存量 clippy 警告（索引循环
改迭代器、冗余闭包、div_ceil、手动 swap、inspect_err、死字段/
死 import 删除、4 处 too_many_arguments 按惯例 allow），CI spice
步骤升级为 -D warnings。

附带：nbody_stm 的 load_kernels() 返回是否加载到 SPK 星历，依赖
MOON/SUN 星历的 4 个测试在无星历时跳过（de430/de440s.bsp 未被
git 跟踪，runner 上没有；本地/服务器有 bsp 时照常真跑），为
spice 的 cargo test 入 CI 做准备。
