+++ .github/workflows/ci.yml
@@ -25,11 +25,9 @@ jobs:
       # Rust 格式与风格
       - run: cargo fmt --all -- --check
       - run: cargo clippy --workspace -- -D warnings
-      # spice-gated 代码（STM 传播、打靶等 Rust 大头）的编译兜底；
+      # spice-gated 代码（STM 传播、打靶等 Rust 大头）的编译与 lint 兜底；
       # cspice-sys 经 downloadcspice 从 NAIF 下载 CSPICE 源码。
-      # 不带 -D warnings：spice-gated 代码从未被 lint 过，存在 30 处预存
-      # clippy 警告（索引循环、冗余闭包等风格类），先保编译覆盖，清零另行处理。
-      - run: cargo clippy --workspace --features spice
+      - run: cargo clippy --workspace --features spice -- -D warnings
       - run: cargo test --workspace
       # Python 格式与风格
       - run: uv sync --group dev
