+++ crates/e2m2e-forces/src/forces/mod.rs
@@ -2,5 +2,9 @@
 //!
 //! 每个 force 一个子模块，1:1 移植自 Python 实现。
 
+pub mod compiled;
+pub mod compiled_stm;
 pub mod gravity_field;
 pub mod nbody_stm;
+pub mod relativistic;
+pub mod srp;
