+++ crates/e2m2e-forces/src/lib.rs
@@ -5,7 +5,7 @@
 #[cfg(feature = "spice")]
 pub mod forces;
 // 纯数学模块，无 spice 依赖；保持无条件 pub，其 #[cfg(test)] 测试在
-// 默认 feature 的 cargo test 下照常执行（integrators 使用自己的本地副本，
-// 不经过本 crate）。
+// 默认 feature 的 cargo test 下照常执行。integrators 的 tide/harmonic 绑定
+// 也直接复用本 crate 的这两个模块。
 pub mod solid_tide;
 pub mod spherical_harmonic;
