+++ crates/e2m2e-integrators/Cargo.toml
@@ -14,6 +14,7 @@ spice = ["dep:cspice", "dep:cspice-sys", "e2m2e-forces/spice", "e2m2e-spice/spic
 [dependencies]
 pyo3 = { workspace = true }
 e2m2e-forces = { path = "../e2m2e-forces" }
+e2m2e-propagation = { path = "../e2m2e-propagation", features = ["pyo3"] }
 e2m2e-spice = { path = "../e2m2e-spice" }
 cspice = { workspace = true, optional = true }
 cspice-sys = { workspace = true, optional = true }
