--- crates/e2m2e-integrators/src/forces/mod.rs
@@ -1,10 +1,0 @@
-//! 力模型 Rust 实现集合（仅 `spice` feature 下编译）。
-//!
-//! 每个 force 一个子模块，1:1 移植自 Python 实现。
-
-pub mod compiled;
-pub mod compiled_stm;
-pub mod gravity_field;
-pub mod nbody_stm;
-pub mod relativistic;
-pub mod srp;
