--- crates/e2m2e-integrators/src/pd78.rs
@@ -1,214 +1,0 @@
-//! Prince-Dormand 8(7)13M 表（Dormand & Prince, 1981；Hairner-Wanner
-//! *Solving Ordinary Differential Equations I*, 2nd ed., p.195-196）。
-//!
-//! 仅系数表存放于此 —— 共享步进逻辑在 [`crate::butcher::explicit_rk_step`]。
-//!
-//! # 来源
-//! 系数逐字转录自 GMAT R2026a
-//! `src/base/propagator/PrinceDormand78.cpp::SetCoefficients()`（第 141 行），
-//! 其中声明 `RungeKutta(13, 8)` —— 13 级，8 阶主解，7 阶嵌入误差控制。
-//! 与 Hairner-Wanner 交叉核对。
-
-use crate::butcher::ButcherTable;
-
-/// Prince-Dormand 8(7)13M：13 级，主阶 8，嵌入阶 7。
-///
-/// 字段映射自 GMAT：`ai`→`c`，`bij`→`a`（严格下三角），
-/// `cj`→`b`（8 阶权重），`cjhat`→`b_star`（7 阶嵌入权重）。
-pub const PD78_TABLE: ButcherTable = ButcherTable::new(
-    13,
-    8,
-    7,
-    // c (GMAT ai)
-    &[
-        0.0,
-        1.0 / 18.0,
-        1.0 / 12.0,
-        0.125,
-        5.0 / 16.0,
-        3.0 / 8.0,
-        59.0 / 400.0,
-        93.0 / 200.0,
-        5490023248.0 / 9719169821.0,
-        13.0 / 20.0,
-        1201146811.0 / 1299019798.0,
-        1.0,
-        1.0,
-    ],
-    // a (GMAT bij, strictly lower triangular)
-    &[
-        &[],
-        &[1.0 / 18.0],
-        &[1.0 / 48.0, 1.0 / 16.0],
-        &[1.0 / 32.0, 0.0, 3.0 / 32.0],
-        &[5.0 / 16.0, 0.0, -75.0 / 64.0, 75.0 / 64.0],
-        &[3.0 / 80.0, 0.0, 0.0, 3.0 / 16.0, 3.0 / 20.0],
-        &[
-            29443841.0 / 614563906.0,
-            0.0,
-            0.0,
-            77736538.0 / 692538347.0,
-            -28693883.0 / 1125000000.0,
-            23124283.0 / 1800000000.0,
-        ],
-        &[
-            16016141.0 / 946692911.0,
-            0.0,
-            0.0,
-            61564180.0 / 158732637.0,
-            22789713.0 / 633445777.0,
-            545815736.0 / 2771057229.0,
-            -180193667.0 / 1043307555.0,
-        ],
-        &[
-            39632708.0 / 573591083.0,
-            0.0,
-            0.0,
-            -433636366.0 / 683701615.0,
-            -421739975.0 / 2616292301.0,
-            100302831.0 / 723423059.0,
-            790204164.0 / 839813087.0,
-            800635310.0 / 3783071287.0,
-        ],
-        &[
-            246121993.0 / 1340847787.0,
-            0.0,
-            0.0,
-            -37695042795.0 / 15268766246.0,
-            -309121744.0 / 1061227803.0,
-            -12992083.0 / 490766935.0,
-            6005943493.0 / 2108947869.0,
-            393006217.0 / 1396673457.0,
-            123872331.0 / 1001029789.0,
-        ],
-        &[
-            -1028468189.0 / 846180014.0,
-            0.0,
-            0.0,
-            8478235783.0 / 508512852.0,
-            1311729495.0 / 1432422823.0,
-            -10304129995.0 / 1701304382.0,
-            -48777925059.0 / 3047939560.0,
-            15336726248.0 / 1032824649.0,
-            -45442868181.0 / 3398467696.0,
-            3065993473.0 / 597172653.0,
-        ],
-        &[
-            185892177.0 / 718116043.0,
-            0.0,
-            0.0,
-            -3185094517.0 / 667107341.0,
-            -477755414.0 / 1098053517.0,
-            -703635378.0 / 230739211.0,
-            5731566787.0 / 1027545527.0,
-            5232866602.0 / 850066563.0,
-            -4093664535.0 / 808688257.0,
-            3962137247.0 / 1805957418.0,
-            65686358.0 / 487910083.0,
-        ],
-        &[
-            403863854.0 / 491063109.0,
-            0.0,
-            0.0,
-            -5068492393.0 / 434740067.0,
-            -411421997.0 / 543043805.0,
-            652783627.0 / 914296604.0,
-            11173962825.0 / 925320556.0,
-            -13158990841.0 / 6184727034.0,
-            3936647629.0 / 1978049680.0,
-            -160528059.0 / 685178525.0,
-            248638103.0 / 1413531060.0,
-            0.0,
-        ],
-    ],
-    // b (GMAT cj): 8th-order weights
-    &[
-        14005451.0 / 335480064.0,
-        0.0,
-        0.0,
-        0.0,
-        0.0,
-        -59238493.0 / 1068277825.0,
-        181606767.0 / 758867731.0,
-        561292985.0 / 797845732.0,
-        -1041891430.0 / 1371343529.0,
-        760417239.0 / 1151165299.0,
-        118820643.0 / 751138087.0,
-        -528747749.0 / 2220607170.0,
-        0.25,
-    ],
-    // b_star (GMAT cjhat): 7th-order embedded weights
-    &[
-        13451932.0 / 455176623.0,
-        0.0,
-        0.0,
-        0.0,
-        0.0,
-        -808719846.0 / 976000145.0,
-        1757004468.0 / 5645159321.0,
-        656045339.0 / 265891186.0,
-        -3867574721.0 / 1518517206.0,
-        465885868.0 / 322736535.0,
-        53011238.0 / 667516719.0,
-        2.0 / 45.0,
-        0.0,
-    ],
-);
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::butcher::explicit_rk_step;
-
-    #[test]
-    fn row_lengths_are_lower_triangular() {
-        for i in 0..PD78_TABLE.stages {
-            assert_eq!(PD78_TABLE.a[i].len(), i, "row {i} should have length {i}");
-        }
-    }
-
-    #[test]
-    fn butcher_rows_sum_to_c_nodes() {
-        for i in 1..PD78_TABLE.stages {
-            let row_sum: f64 = PD78_TABLE.a[i].iter().sum();
-            assert!(
-                (row_sum - PD78_TABLE.c[i]).abs() < 1e-12,
-                "row {i} sum {row_sum} != c {}",
-                PD78_TABLE.c[i]
-            );
-        }
-    }
-
-    #[test]
-    fn weights_sum_to_one() {
-        let b_sum: f64 = PD78_TABLE.b.iter().sum();
-        let b_star_sum: f64 = PD78_TABLE.b_star.iter().sum();
-        assert!((b_sum - 1.0).abs() < 1e-12, "b sum {b_sum} != 1");
-        assert!((b_star_sum - 1.0).abs() < 1e-12, "b* sum {b_star_sum} != 1");
-    }
-
-    #[test]
-    fn harmonic_oscillator_small_step_error() {
-        // y'' = -y as first-order system [y, v]: y' = v, v' = -y.
-        let f = |_t: f64, state: &[f64]| -> Result<Vec<f64>, std::convert::Infallible> {
-            Ok(vec![state[1], -state[0]])
-        };
-
-        let y0 = vec![1.0, 0.0];
-        let h = 1e-4;
-
-        let (y1, error) = explicit_rk_step(&PD78_TABLE, 0.0, &y0, h, f, None).unwrap();
-
-        let y_exact = vec![h.cos(), -h.sin()];
-        let num_err = y1
-            .iter()
-            .zip(y_exact.iter())
-            .map(|(a, b)| (a - b).powi(2))
-            .sum::<f64>()
-            .sqrt();
-
-        // 8th-order method: far more accurate than PD45 at the same step.
-        assert!(num_err < 1e-12, "numerical error {num_err} too large");
-        assert!(error < 1e-12, "local truncation estimate {error} too large");
-    }
-}
